Explore Difference Between Var, Let & Const
One of the biggest problems with declaring variables with the var keyword is that you can easily overwrite variable declarations:

A keyword called let & const was introduced in ES6, a major update to JavaScript, to solve this potential issue with the var keyword. You'll learn about other ES6 features in later challenges.